More testing - autouse=True is problematic
