+++ tests/e2e/alpha/internal/composition/test_attributes.py
@@ -25,7 +25,6 @@ FLOAT_SERIES_VALUES = [float(step**2) for step in range(10)]
 
 @pytest.fixture(scope="module")
 def run_with_attributes(client, project):
-    raise Exception("test_attributes.py:run_with_attributes")
     import uuid
 
     from neptune_scale import Run
